+++ fltk/src/misc.rs
@@ -536,7 +536,7 @@ impl InputChoice {
         }
     }
 
-    
+    /// Add an element to the input choice
     pub fn add(&mut self, s: &str) {
         assert!(!self.was_deleted());
         let s = CString::safe_new(s);
@@ -545,7 +545,7 @@ impl InputChoice {
         }
     }
 
-
+    /// Clear the input choice widget
     pub fn clear(&mut self) {
         assert!(!self.was_deleted());
         unsafe {
@@ -553,7 +553,7 @@ impl InputChoice {
         }
     }
 
-
+    /// Get the value of the current choice
     pub fn value(&self) -> Option<String> {
         assert!(!self.was_deleted());
         unsafe {
@@ -566,7 +566,7 @@ impl InputChoice {
         }
     }
 
-
+    /// Set the value to a string
     pub fn set_value(&mut self, val: &str) {
         assert!(!self.was_deleted());
         let val = CString::safe_new(val);
@@ -575,16 +575,20 @@ impl InputChoice {
         }
     }
 
-
-    pub fn set_value2(&mut self, val: i32) {
+    /// Set the value of the input choice to a current element
+    pub fn set_value2(&mut self, val: u32) {
+        debug_assert!(
+            val <= std::isize::MAX as u32,
+            "u32 entries have to be < std::isize::MAX for compatibility!"
+        );
         assert!(!self.was_deleted());
         unsafe {
-            Fl_Input_Choice_set_value2(self._inner, val)
+            Fl_Input_Choice_set_value2(self._inner, val as i32)
         }
     }
 
-
-    pub fn menubutton(&mut self) -> Box<dyn MenuExt> {
+    /// Get the associated menu button
+    pub fn menu_button(&mut self) -> Box<dyn MenuExt> {
         assert!(!self.was_deleted());
         unsafe {
             let ptr = Fl_Input_Choice_menubutton(self._inner);
@@ -642,6 +646,7 @@ pub struct HelpView {
 }
 
 impl HelpView {
+    /// Return the directory
     pub fn directory(&self) -> std::path::PathBuf {
         assert!(!self.was_deleted());
         unsafe {
@@ -658,7 +663,7 @@ impl HelpView {
         }
     }
 
-
+    /// Return the filename
     pub fn filename(&self) -> std::path::PathBuf {
         assert!(!self.was_deleted());
         unsafe {
@@ -675,16 +680,20 @@ impl HelpView {
         }
     }
 
-
-    pub fn find(&mut self, s: &str, p: usize) -> i32 {
+    /// Find a string, returns the index
+    pub fn find(&mut self, s: &str, start_from: u32) -> u32 {
         assert!(!self.was_deleted());
+        debug_assert!(
+            start_from <= std::isize::MAX as u32,
+            "u32 entries have to be < std::isize::MAX for compatibility!"
+        );
         unsafe {
             let s = CString::safe_new(s);
-            Fl_Help_View_find(self._inner, s.as_ptr(), p as i32)
+            Fl_Help_View_find(self._inner, s.as_ptr(), p as i32) as u32
         }
     }
 
-
+    /// Get the value of the widget
     pub fn value(&self) -> Option<String> {
         assert!(!self.was_deleted());
         unsafe {
@@ -697,51 +706,51 @@ impl HelpView {
         }
     }
 
-
+    /// Set value of the widget
     pub fn set_value(&mut self, val: &str) {
         assert!(!self.was_deleted());
         let val = CString::safe_new(val);
         unsafe { Fl_Help_View_set_value(self._inner, val.as_ptr()) }
     }
 
-
+    /// Clear selection
     pub fn clear_selection(&mut self) {
         assert!(!self.was_deleted());
         unsafe { Fl_Help_View_clear_selection(self._inner) }
     }
 
-
+    /// Select all
     pub fn select_all(&mut self) {
         assert!(!self.was_deleted());
         unsafe { Fl_Help_View_select_all(self._inner) }
     }
 
-
+    /// Set the topline string
     pub fn set_topline(&mut self, n: &str) {
         assert!(!self.was_deleted());
         let n = CString::safe_new(n);
         unsafe { Fl_Help_View_set_topline(self._inner, n.as_ptr()) }
     }
 
-
+    /// Set the leftline position
     pub fn set_topline2(&mut self, arg1: i32) {
         assert!(!self.was_deleted());
         unsafe { Fl_Help_View_set_topline2(self._inner, arg1) }
     }
 
-
+    /// Get the topline position
     pub fn topline(&self) -> i32 {
         assert!(!self.was_deleted());
         unsafe { Fl_Help_View_topline(self._inner) }
     }
 
-
+    /// Set the leftline position
     pub fn set_leftline(&mut self, arg1: i32) {
         assert!(!self.was_deleted());
         unsafe { Fl_Help_View_set_leftline(self._inner, arg1) }
     }
 
-
+    /// Gets the current leftline in pixels
     pub fn leftline(&self) -> i32 {
         assert!(!self.was_deleted());
         unsafe { Fl_Help_View_leftline(self._inner) }
@@ -805,237 +814,10 @@ impl HelpView {
         unsafe { Fl_Help_View_set_scrollbar_size(self._inner, new_size as i32) }
     }
 
-
-    pub fn load(&mut self, f: &str) -> i32 {
+    /// Load a view from a file or URI
+    pub fn load<P: AsRef<std::path::Path>>(&mut self, f: P) -> i32 {
         assert!(!self.was_deleted());
-        let f = CString::safe_new(f);
+        let f = CString::safe_new(f.as_ref());
         unsafe { Fl_Help_View_load(self._inner, f.as_ptr()) }
-    }
-}
-
-/// Creates a CheckBrowser widget
-#[derive(WidgetBase, WidgetExt, Debug)]
-pub struct CheckBrowser {
-    _inner: *mut Fl_Check_Browser,
-    _tracker: *mut fltk_sys::fl::Fl_Widget_Tracker,
-}
-
-impl CheckBrowser {
-    
-    pub fn add(&mut self, s: &str, b: bool) -> i32 {
-        assert!(!self.was_deleted());
-        let s = CString::safe_new(s);
-        unsafe {
-            Fl_Check_Browser_add(self._inner, s.as_ptr(), b as i32)
-        }
-    }
-
-
-    pub fn remove(&mut self, item: usize) -> i32 {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_remove(self._inner, item as i32)
-        }
-    }
-
-
-    pub fn clear(&mut self) {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_clear(self._inner)
-        }
-    }
-
-
-    pub fn nitems(&self) -> usize {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_nitems(self._inner) as usize
-        }
-    }
-
-
-    pub fn nchecked(&self) -> usize {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_nchecked(self._inner) as usize
-        }
-    }
-
-
-    pub fn checked(&self, item: i32) -> bool {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_checked(self._inner, item) != 0
-        }
-    }
-
-
-    pub fn set_checked(&mut self, item: i32) {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_set_checked(self._inner, item)
-        }
-    }
-
-
-    pub fn check_all(&mut self) {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_check_all(self._inner)
-        }
-    }
-
-
-    pub fn check_none(&mut self) {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_check_none(self._inner)
-        }
-    }
-
-
-    pub fn value(&self) -> i32 {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_value(self._inner)
-        }
-    }
-
-
-    pub fn text(&self, item: i32) -> Option<String> {
-        assert!(!self.was_deleted());
-        unsafe {
-            let ptr = Fl_Check_Browser_text(self._inner, item);
-            if ptr.is_null() {
-                None
-            } else {
-                Some(CStr::from_ptr(ptr).to_string_lossy().to_string())
-            }
-        }
-    }
-
-    /// Gets the text font
-    pub fn text_font(&self) -> Font {
-        assert!(!self.was_deleted());
-        unsafe { std::mem::transmute(Fl_Check_Browser_textfont(self._inner)) }
-    }
-
-    /// Sets the text font
-    pub fn set_text_font(&mut self, f: Font) {
-        assert!(!self.was_deleted());
-        unsafe { Fl_Check_Browser_set_textfont(self._inner, f.bits() as i32) }
-    }
-
-    /// Gets the text size
-    pub fn text_size(&self) -> u32 {
-        assert!(!self.was_deleted());
-        unsafe { Fl_Check_Browser_textsize(self._inner) as u32 }
-    }
-
-    /// Sets the text size
-    pub fn set_text_size(&mut self, s: u32) {
-        debug_assert!(
-            s <= std::isize::MAX as u32,
-            "u32 entries have to be < std::isize::MAX for compatibility!"
-        );
-        assert!(!self.was_deleted());
-        unsafe { Fl_Check_Browser_set_textsize(self._inner, s as i32) }
-    }
-
-    /// Gets the text's color
-    pub fn text_color(&self) -> Color {
-        assert!(!self.was_deleted());
-        unsafe { std::mem::transmute(Fl_Check_Browser_textcolor(self._inner)) }
-    }
-
-    /// Sets the text's color
-    pub fn set_text_color(&mut self, color: Color) {
-        assert!(!self.was_deleted());
-        unsafe { Fl_Check_Browser_set_textcolor(self._inner, color.bits() as u32) }
-    }
-
-    pub fn position(&self) -> u32 {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_position(self._inner) as u32
-        }
-    }
-
-    pub fn set_position(&mut self, pos: u32) {
-        assert!(!self.was_deleted());
-        debug_assert!(pos <= std::isize::MAX as u32, "u32 entries have to be < std::isize::MAX for compatibility!");
-        unsafe {
-            Fl_Check_Browser_set_position(self._inner, pos as i32)
-        }
-    }
-
-    pub fn hposition(&self) -> u32 {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_hposition(self._inner) as u32
-        }
-    }
-
-    pub fn set_hposition(&mut self, pos: u32) {
-        assert!(!self.was_deleted());
-        debug_assert!(pos <= std::isize::MAX as u32, "u32 entries have to be < std::isize::MAX for compatibility!");
-        unsafe {
-            Fl_Check_Browser_set_hposition(self._inner, pos as i32)
-        }
-    }
-
-    pub fn has_scrollbar(&self) -> crate::browser::BrowserScrollbar {
-        assert!(!self.was_deleted());
-        unsafe {
-            mem::transmute(Fl_Check_Browser_has_scrollbar(self._inner))
-        }
-    }
-
-    pub fn set_has_scrollbar(&mut self, mode: crate::browser::BrowserScrollbar) {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_set_has_scrollbar(self._inner, mode as raw::c_uchar)
-        }
-    }
-
-    pub fn scrollbar_size(&self) -> u32 {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_scrollbar_size(self._inner) as u32
-        }
-    }
-
-    pub fn set_scrollbar_size(&mut self, new_size: u32) {
-        assert!(!self.was_deleted());
-        debug_assert!(new_size <= std::isize::MAX as u32, "u32 entries have to be < std::isize::MAX for compatibility!");
-        unsafe {
-            Fl_Check_Browser_set_scrollbar_size(self._inner, new_size as i32)
-        }
-    }
-
-    pub fn sort(&mut self) {
-        assert!(!self.was_deleted());
-        unsafe {
-            Fl_Check_Browser_sort(self._inner)
-        }
-    }
-
-    pub fn scrollbar(&self) -> Box<dyn ValuatorExt> {
-        assert!(!self.was_deleted());
-        unsafe {
-            let ptr = Fl_Check_Browser_scrollbar(self._inner);
-            assert!(!ptr.is_null());
-            Box::new(crate::valuator::Scrollbar::from_widget_ptr(ptr as *mut fltk_sys::widget::Fl_Widget))
-        }
-    }
-
-    pub fn hscrollbar(&self) -> Box<dyn ValuatorExt> {
-        assert!(!self.was_deleted());
-        unsafe {
-            let ptr = Fl_Check_Browser_hscrollbar(self._inner);
-            assert!(!ptr.is_null());
-            Box::new(crate::valuator::Scrollbar::from_widget_ptr(ptr as *mut fltk_sys::widget::Fl_Widget))
-        }
     }
 }